requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
attrs==21.4.0
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cymem==2.0.5
docutils==0.19
Flask==2.0.2
Flask-Cors==3.0.10
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.1.0/fr_core_news_md-3.1.0-py3-none-any.whl
future==0.18.2
idna==3.2
iso8601==1.0.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.4.0
Kivy==2.1.0
Kivy-Garden==0.1.5
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.2
p2pnetwork==1.1
packaging==21.0
pathy==0.6.0
Pillow==8.3.2
preshed==3.0.5
pydantic==1.8.2
pygame==2.1.2
Pygments==2.12.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-telegram-bot==13.7
pytz==2021.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
telegram==0.0.1
thinc==8.0.10
tinydb==4.5.1
tornado==6.1
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.6
wasabi==0.8.2
watchdog==2.1.9
Werkzeug==2.0.2
